Very minor copy edits
Just fyi…I had removed your acronyms for TPP and VaaS because we’re not supposed to use them. We’re asked to always spell these out. In Doc, we use variables, which makes this really easy. Not sure you can do that in github. So I left them for now.